+++ applications/amathea/src/features/map/Map.tsx
@@ -4,25 +4,29 @@ import { InteractiveMap, MapRequest } from '@globalfishingwatch/react-map-gl'
 import Miniglobe, { MiniglobeBounds } from '@globalfishingwatch/ui-components/dist/miniglobe'
 import IconButton from '@globalfishingwatch/ui-components/dist/icon-button'
 import useLayerComposer from '@globalfishingwatch/react-hooks/dist/use-layer-composer'
-import GFWAPI from '@globalfishingwatch/api-client'
 import { useAOIConnect } from 'features/areas-of-interest/areas-of-interest.hook'
+import { useUserConnect } from 'features/user/user.hook'
 import { useGeneratorsConnect, useViewport } from './map.hooks'
 import styles from './Map.module.css'
 
 import '@globalfishingwatch/mapbox-gl/dist/mapbox-gl.css'
+
+const mapOptions = {
+  customAttribution: '© Copyright Global Fishing Watch 2020',
+}
 
 const Map = (): React.ReactElement => {
   const mapRef = useRef<any>(null)
   const { viewport, onViewportChange, setMapCoordinates } = useViewport()
   const { latitude, longitude, zoom } = viewport
   const { currentAOI } = useAOIConnect()
+  const { logged, token } = useUserConnect()
   const { generatorsConfig, globalConfig } = useGeneratorsConnect()
   // useLayerComposer is a convenience hook to easily generate a Mapbox GL style (see https://docs.mapbox.com/mapbox-gl-js/style-spec/) from
   // the generatorsConfig (ie the map "layers") and the global configuration
   const { style } = useLayerComposer(generatorsConfig, globalConfig)
-  const token = GFWAPI.getToken()
 
-  const [bounds, setBounds] = useState<MiniglobeBounds | null>(null)
+  const [bounds, setBounds] = useState<MiniglobeBounds | undefined>()
 
   const transformRequest: any = useCallback(
     (url: string, resourceType: string) => {
@@ -73,9 +77,16 @@ const Map = (): React.ReactElement => {
     }
   }, [currentAOI, setMapCoordinates])
 
+  const onZoomInClick = useCallback(() => {
+    setMapCoordinates({ latitude, longitude, zoom: zoom + 1 })
+  }, [latitude, longitude, setMapCoordinates, zoom])
+  const onZoomOutClick = useCallback(() => {
+    setMapCoordinates({ latitude, longitude, zoom: Math.max(1, zoom - 1) })
+  }, [latitude, longitude, setMapCoordinates, zoom])
+
   return (
     <div className={styles.container}>
-      {style && (
+      {style && logged && token && (
         <InteractiveMap
           ref={mapRef}
           width="100%"
@@ -86,31 +97,15 @@ const Map = (): React.ReactElement => {
           zoom={zoom}
           onViewportChange={onViewportChange}
           mapStyle={style}
-          mapOptions={{
-            customAttribution: '© Copyright Global Fishing Watch 2020',
-          }}
+          mapOptions={mapOptions}
         />
       )}
       <div className={styles.mapControls}>
-        {bounds && <Miniglobe size={60} bounds={bounds} center={{ latitude, longitude }} />}
-        <IconButton
-          icon="plus"
-          type="map-tool"
-          tooltip="Zoom in"
-          onClick={() => {
-            setMapCoordinates({ latitude, longitude, zoom: zoom + 1 })
-          }}
-        />
-        <IconButton
-          icon="minus"
-          type="map-tool"
-          tooltip="Zoom out"
-          onClick={() => {
-            setMapCoordinates({ latitude, longitude, zoom: Math.max(1, zoom - 1) })
-          }}
-        />
-        <IconButton icon="ruler" type="map-tool" tooltip="Open ruler tool" />
-        <IconButton icon="camera" type="map-tool" tooltip="Capture the map" />
+        <Miniglobe size={60} bounds={bounds} center={{ latitude, longitude }} />
+        <IconButton icon="plus" type="map-tool" tooltip="Zoom in" onClick={onZoomInClick} />
+        <IconButton icon="minus" type="map-tool" tooltip="Zoom out" onClick={onZoomOutClick} />
+        <IconButton icon="ruler" type="map-tool" tooltip="Open ruler tool (Coming soon)" />
+        <IconButton icon="camera" type="map-tool" tooltip="Capture the map (Coming soon)" />
       </div>
     </div>
   )