+++ applications/amathea/src/features/map/map.slice.ts
@@ -6,6 +6,7 @@ import {
   BackgroundGeneratorConfig,
   BasemapGeneratorConfig,
 } from '@globalfishingwatch/layer-composer/dist/generators/types'
+import GFWAPI from '@globalfishingwatch/api-client'
 import { selectMapZoomQuery, selectTimerange } from 'routes/routes.selectors'
 
 export interface MapState {
@@ -20,7 +21,7 @@ const initialState: MapState = {
       type: Generators.Type.Background,
       color: '#002458',
     } as BackgroundGeneratorConfig,
-    { id: 'satellite', type: Generators.Type.Basemap, visible: true } as BasemapGeneratorConfig,
+    { id: 'landmass', type: Generators.Type.Basemap, visible: true } as BasemapGeneratorConfig,
   ],
 }
 
@@ -57,6 +58,7 @@ export const selectGlobalGeneratorsConfig = createSelector(
     zoom,
     start,
     end,
+    token: GFWAPI.getToken(),
   })
 )