+++ packages/layer-composer/src/generators/background/background.ts
@@ -1,21 +1,22 @@
 import { Layer } from 'mapbox-gl'
 import { Group } from '../../types'
 import { Type, BackgroundGeneratorConfig } from '../types'
+import { isConfigVisible } from '../utils'
 
 export const DEFAULT_BASEMAP_COLOR = '#002457'
 
 class BackgroundGenerator {
   type = Type.Background
 
-  _getStyleLayers = (layer: BackgroundGeneratorConfig): Layer[] => [
+  _getStyleLayers = (config: BackgroundGeneratorConfig): Layer[] => [
     {
       id: 'background',
       type: 'background',
       layout: {
-        visibility: layer.visible !== undefined ? (layer.visible ? 'visible' : 'none') : 'visible',
+        visibility: isConfigVisible(config),
       },
       paint: {
-        'background-color': layer.color || DEFAULT_BASEMAP_COLOR,
+        'background-color': config.color || DEFAULT_BASEMAP_COLOR,
       },
       metadata: {
         group: Group.Background,