+++ packages/layer-composer/src/generators/types.ts
@@ -1,8 +1,10 @@
 import { FeatureCollection } from 'geojson'
 import { Segment } from './track/segments-to-geojson'
+import { Geoms } from './heatmap/types'
 
 export enum Type {
   Background = 'BACKGROUND',
+  UserContext = 'USER_CONTEXT',
   Basemap = 'BASEMAP',
   CartoPolygons = 'CARTO_POLYGONS',
   GL = 'GL',
@@ -18,6 +20,7 @@ export interface GlobalGeneratorConfig {
   end: string
   zoom: number
   zoomLoadLevel?: number
+  token?: string
 }
 
 export type AnyData = FeatureCollection | Segment[] | RawEvent[] | Ruler[]
@@ -56,6 +59,21 @@ export interface BackgroundGeneratorConfig extends GeneratorConfig {
    * Sets the color of the map background in any format supported by Mapbox GL, see https://docs.mapbox.com/mapbox-gl-js/style-spec/types/#color
    */
   color?: string
+}
+
+/**
+ * Layers created by user uploading their own shapefile
+ */
+export interface UserContextGeneratorConfig extends GeneratorConfig {
+  type: Type.UserContext
+  /**
+   * Sets the color of the line https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/#paint-fill-fill-color
+   */
+  color?: string
+  /**
+   * Url to grab the tiles from, internally using https://docs.mapbox.com/mapbox-gl-js/style-spec/sources/#vector-tiles
+   */
+  tilesUrl: string
 }
 
 /**
@@ -139,9 +157,11 @@ export interface HeatmapGeneratorConfig extends GeneratorConfig {
   // end: string
   // zoom: number
   maxZoom?: number
-  tileset: string
+  tilesUrl: string
+  statsUrl?: string
   fetchStats?: boolean
   statsFilter?: string
+  geomType?: Geoms
   colorRamp?: ColorRampsIds
   serverSideFilter?: string
   updateColorRampOnTimeChange?: boolean
@@ -154,7 +174,7 @@ export interface HeatmapAnimatedGeneratorConfig extends GeneratorConfig {
   colorRamps?: ColorRampsIds[]
   tilesAPI?: string
   combinationMode?: CombinationMode
-  geomType?: string
+  geomType?: Geoms
   tilesetStart?: string
   tilesetEnd?: string
   maxZoom?: number
@@ -169,6 +189,7 @@ export type AnyGeneratorConfig =
   | BasemapGeneratorConfig
   | GlGeneratorConfig
   | CartoPolygonsGeneratorConfig
+  | UserContextGeneratorConfig
   | TrackGeneratorConfig
   | VesselEventsGeneratorConfig
   | RulersGeneratorConfig
@@ -211,7 +232,7 @@ export type Ruler = {
   isNew?: boolean
 }
 
-// ---- Heatmap Generator types
+// ---- Heatmap Generator color ramps types
 export type ColorRampsIds =
   | 'fishing'
   | 'presence'