+++ packages/react-hooks/src/use-layer-composer/types.d.ts
@@ -1,2 +1,2 @@
 declare module '@mapbox/mapbox-gl-style-spec'
-declare module 'mapbox-gl/dist/style-spec'
+declare module '@globalfishingwatch/mapbox-gl/dist/style-spec'