+++ applications/sandbox/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.2.1
+
+### Patch Changes
+
+- 152d7ed: y
+
+## 0.2.0
+
+### Minor Changes
+
+- 0413cc6: Memoize functional components
+
 ## 0.1.14
 
 ### Patch Changes